+++ Drivers/SPI/spi.h
@@ -7,7 +7,7 @@
 #define SPI_QUEUE_SIZE		(4)
 
 void	 SPI0_Init(void);
-bool     SPI_SelfTest(Spi *spi, uint16_t *msg, uint16_t *recv, uint32_t len);
+bool     SPI_SelfTest(Spi *spi, uint8_t *msg, uint8_t *recv, uint32_t len);
 uint16_t SPI0_vTransmitHalfword(uint16_t const halfword);
 uint8_t  SPI0_vTransmitByte(uint8_t const byte);
